+++ src/a04_reason_logic/test/test_fact.py
@@ -1,5 +1,5 @@
 from src.a01_term_logic.rope import create_rope, get_default_central_label as root_label
-from src.a04_reason_logic.reason_plan import (
+from src.a04_reason_logic.reason import (
     FactCore,
     FactUnit,
     factheir_shop,
