+++ src/a05_plan_logic/test/test_plan/test_plan_dict.py
@@ -1,7 +1,7 @@
 from src.a01_term_logic.rope import create_rope
 from src.a03_group_logic.group import awardunit_shop
 from src.a03_group_logic.labor import laborunit_shop
-from src.a04_reason_logic.reason_plan import (
+from src.a04_reason_logic.reason import (
     caseunit_shop,
     factunit_shop,
     reasonheir_shop,
