+++ src/a05_plan_logic/test/test_plan/test_plan_subsitute_rope.py
@@ -1,9 +1,5 @@
 from src.a01_term_logic.rope import create_rope
-from src.a04_reason_logic.reason_plan import (
-    caseunit_shop,
-    factunit_shop,
-    reasonunit_shop,
-)
+from src.a04_reason_logic.reason import caseunit_shop, factunit_shop, reasonunit_shop
 from src.a05_plan_logic.plan import planunit_shop
