+++ src/a06_belief_logic/test/_util/example_beliefs.py
@@ -1,7 +1,7 @@
 from src.a00_data_toolbox.file_toolbox import open_file
 from src.a01_term_logic.rope import RopeTerm
 from src.a03_group_logic.labor import laborunit_shop
-from src.a04_reason_logic.reason_plan import factunit_shop, reasonunit_shop
+from src.a04_reason_logic.reason import factunit_shop, reasonunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import (
     BeliefUnit,
