+++ src/a06_belief_logic/test/test_belief/test_belief_json.py
@@ -3,7 +3,7 @@ from src.a00_data_toolbox.dict_toolbox import get_dict_from_json, x_is_json
 from src.a01_term_logic.rope import default_knot_if_None, to_rope
 from src.a03_group_logic.group import awardunit_shop
 from src.a03_group_logic.labor import laborunit_shop, partyunit_shop
-from src.a04_reason_logic.reason_plan import factunit_shop
+from src.a04_reason_logic.reason import factunit_shop
 from src.a05_plan_logic.healer import healerunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import (
