+++ src/a06_belief_logic/test/test_belief/test_belief_relevant_ropes.py
@@ -1,5 +1,5 @@
 from src.a01_term_logic.rope import to_rope
-from src.a04_reason_logic.reason_plan import reasonunit_shop
+from src.a04_reason_logic.reason import reasonunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a06_belief_logic.test._util.example_beliefs import (
