+++ src/a06_belief_logic/test/test_belief_cash_out/test_init_plan_tree_walk.py
@@ -1,5 +1,5 @@
 from src.a01_term_logic.rope import to_rope
-from src.a04_reason_logic.reason_plan import reasonunit_shop
+from src.a04_reason_logic.reason import reasonunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop, get_sorted_plan_list
 from src.a06_belief_logic.test._util.example_beliefs import get_beliefunit_with_4_levels
