+++ src/a06_belief_logic/test/test_belief_cash_out/test_tree_traverse_z_chore_status.py
@@ -1,10 +1,6 @@
 from src.a01_term_logic.rope import to_rope
 from src.a02_finance_logic.finance_config import default_fund_pool
-from src.a04_reason_logic.reason_plan import (
-    caseunit_shop,
-    reasonheir_shop,
-    reasonunit_shop,
-)
+from src.a04_reason_logic.reason import caseunit_shop, reasonheir_shop, reasonunit_shop
 from src.a05_plan_logic.healer import healerunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_graphics import display_plantree
