+++ src/a06_belief_logic/test/test_belief_cash_out/test_tree_traverse_z_fact.py
@@ -1,5 +1,5 @@
 from pytest import raises as pytest_raises
-from src.a04_reason_logic.reason_plan import factheir_shop, factunit_shop
+from src.a04_reason_logic.reason import factheir_shop, factunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a06_belief_logic.test._util.example_beliefs import (
