+++ src/a06_belief_logic/test/test_belief_cash_out/test_z_set_knot.py
@@ -1,6 +1,6 @@
 from pytest import raises as pytest_raises
 from src.a01_term_logic.rope import create_rope, to_rope
-from src.a04_reason_logic.reason_plan import factunit_shop, reasonunit_shop
+from src.a04_reason_logic.reason import factunit_shop, reasonunit_shop
 from src.a05_plan_logic.plan import (
     get_default_moment_label as root_label,
     planunit_shop,
