+++ src/a06_belief_logic/tree_metrics.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from src.a00_data_toolbox.dict_toolbox import get_0_if_None, get_empty_dict_if_None
 from src.a01_term_logic.term import GroupTitle
 from src.a03_group_logic.group import AwardUnit
-from src.a04_reason_logic.reason_plan import ReasonUnit, RopeTerm
+from src.a04_reason_logic.reason import ReasonUnit, RopeTerm
 
 
 @dataclass
