+++ src/a07_timeline_logic/test/test_/test_config_creg_.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from src.a00_data_toolbox.plotly_toolbox import conditional_fig_show
 from src.a03_group_logic.group import awardunit_shop
-from src.a04_reason_logic.reason_plan import reasonunit_shop
+from src.a04_reason_logic.reason import reasonunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a07_timeline_logic.test._util.a07_str import (
