+++ src/a07_timeline_logic/test/test_timeline_str/test_timeline_str_func.py
@@ -3,7 +3,7 @@ from src.a01_term_logic.rope import (
     default_knot_if_None,
     get_default_central_label as root_label,
 )
-from src.a04_reason_logic.reason_plan import factunit_shop, reasonunit_shop
+from src.a04_reason_logic.reason import factunit_shop, reasonunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a07_timeline_logic.reason_str_func import (
     get_fact_state_readable_str,
