+++ src/a08_belief_atom_logic/test/test_atom/test_atom_sift_update.py
@@ -1,5 +1,5 @@
 from src.a03_group_logic.group import awardunit_shop
-from src.a04_reason_logic.reason_plan import factunit_shop, reasonunit_shop
+from src.a04_reason_logic.reason import factunit_shop, reasonunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a06_belief_logic.belief_tool import (
     belief_plan_factunit_get_obj,
