+++ src/a09_pack_logic/test/test_delta/test_delta_get_belief_difference.py
@@ -5,7 +5,7 @@ from src.a00_data_toolbox.dict_toolbox import (
 )
 from src.a03_group_logic.group import awardunit_shop
 from src.a03_group_logic.partner import partnerunit_shop
-from src.a04_reason_logic.reason_plan import factunit_shop
+from src.a04_reason_logic.reason import factunit_shop
 from src.a05_plan_logic.plan import planunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a06_belief_logic.test._util.a06_str import (
