+++ src/a11_bud_logic/test/test_cell_.py
@@ -1,6 +1,6 @@
 from copy import deepcopy as copy_deepcopy
 from src.a01_term_logic.rope import create_rope
-from src.a04_reason_logic.reason_plan import factunit_shop
+from src.a04_reason_logic.reason import factunit_shop
 from src.a06_belief_logic.belief_main import beliefunit_shop
 from src.a11_bud_logic.cell import (
     CELLNODE_QUOTA_DEFAULT,
