+++ src/a12_hub_toolbox/fact_tool.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.dict_toolbox import get_max_key
 from src.a01_term_logic.term import BeliefName, RopeTerm
-from src.a04_reason_logic.reason_plan import FactUnit, get_factunit_from_tuple
+from src.a04_reason_logic.reason import FactUnit, get_factunit_from_tuple
 
 
 def get_nodes_with_weighted_facts(
